// src/generated/models/Cre2b_personnelsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cre2b_personnelscre2b_role = {
  200080000: 'TSR',
  200080001: 'Advisor',
  200080002: 'Role'
} as const;
export type Cre2b_personnelscre2b_role = keyof typeof Cre2b_personnelscre2b_role;
export const Cre2b_personnelscre2b_status = {
  200080000: 'Available',
  200080001: 'Unavailable'
} as const;
export type Cre2b_personnelscre2b_status = keyof typeof Cre2b_personnelscre2b_status;
export const Cre2b_personnelsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cre2b_personnelsstatecode = keyof typeof Cre2b_personnelsstatecode;
export const Cre2b_personnelsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cre2b_personnelsstatuscode = keyof typeof Cre2b_personnelsstatuscode;

export interface Cre2b_personnelsBase {
  cre2b_fullname?: string;
  cre2b_manager?: string;
  cre2b_personnelid: string;
  cre2b_personnelid1?: string;
  cre2b_role?: Cre2b_personnelscre2b_role;
  cre2b_status?: Cre2b_personnelscre2b_status;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Cre2b_personnelsstatecode;
  statuscode?: Cre2b_personnelsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cre2b_personnels extends Cre2b_personnelsBase {
  cre2b_rolename?: string;
  cre2b_statusname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}